Feature: Call parent events from subclass obs
This change creates a bit more of boilerplate (since we are storing some
data probably more than once, eg we have 2 subclasses from a parent and
the parent has events, that method is stored twice). But its the same
pointer so its not that harsh. On the other hand, this gives us the
advantage to being able to call the parent avoiding another for (+ TODO
Check if I override the parent class if it wont duplicate the method or
whom it will be called)